requirements.txt
PyYAML==5.3.1
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
future==0.18.2
idna==2.10
loguru==0.5.3
mimesis==4.1.2
pefile==2019.4.18	
pip==20.2.3
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pystun3==1.0.0
pywin32-ctypes==0.2.0
requests==2.24.0
selenium==3.141.0
setuptools==50.3.1	
urllib3==1.25.10
win32-setctime==1.0.3